I changed the odd home container to jobs container. I created a new home container that is responsive and I added and change the route in the App.tsx.
